style: lint contracts

+++ contracts/CompoundModule.sol
@@ -78,14 +78,16 @@ contract CompoundModule {
         lendingBalanceOf[msg.sender].unused += _amount; // In cToken.
     }
 
-    event MyLog(string, uint);
+    event MyLog(string, uint256);
 
     /** @dev Allows someone to borrow ETH.)
      *  @param _amount Amount to borrow in ETH.
      */
     function borrow(uint256 _amount) external {
         // Calculate the collateral required.
-        uint256 daiAmountEquivalentToEthAmount = _amount.mul(1e18).div(oracle.consult());
+        uint256 daiAmountEquivalentToEthAmount = _amount.mul(1e18).div(
+            oracle.consult()
+        );
         uint256 collateralRequiredInDai = daiAmountEquivalentToEthAmount
         .mul(collateralFactor)
         .div(1e18);
@@ -213,7 +215,9 @@ contract CompoundModule {
             amountInCDai <= collateralBalanceOf[msg.sender],
             "Must redeem less than collateral."
         );
-        uint256 borrowingAmount = borrowingBalanceOf[msg.sender].mul(1e18).div(oracle.consult());
+        uint256 borrowingAmount = borrowingBalanceOf[msg.sender].mul(1e18).div(
+            oracle.consult()
+        );
         uint256 borrowingAmountInDai = borrowingAmount;
         uint256 collateralAfterInCDAI = collateralBalanceOf[msg.sender].sub(
             amountInCDai
@@ -243,7 +247,9 @@ contract CompoundModule {
         );
         _payBack(_borrower, msg.value);
         uint256 daiToEthRate = oracle.consult();
-        uint256 borrowingAmountInDai = borrowingBalanceOf[_borrower].mul(1e18).div(daiToEthRate);
+        uint256 borrowingAmountInDai = borrowingBalanceOf[_borrower]
+        .mul(1e18)
+        .div(daiToEthRate);
         uint256 repayAmountInDai = msg.value.mul(1e18).div(daiToEthRate);
         uint256 daiExchangeRate = cDaiToken.exchangeRateCurrent();
         uint256 collateralInDai = collateralBalanceOf[_borrower]
@@ -281,7 +287,9 @@ contract CompoundModule {
         public
         returns (uint256)
     {
-        uint256 borrowingAmountInDai = borrowingBalanceOf[_borrower].mul(1e18).div(oracle.consult());
+        uint256 borrowingAmountInDai = borrowingBalanceOf[_borrower]
+        .mul(1e18)
+        .div(oracle.consult());
         uint256 collateralRequiredInDai = borrowingAmountInDai
         .mul(collateralFactor)
         .div(1e18);
