+++ contracts/oracle/ChainlinkOracle.sol
@@ -3,7 +3,6 @@ pragma solidity >=0.6.6;
 import "@chainlink/contracts/src/v0.6/interfaces/AggregatorV3Interface.sol";
 
 contract ChainlinkOracle {
-
     AggregatorV3Interface internal priceFeed;
 
     /**
@@ -12,18 +11,20 @@ contract ChainlinkOracle {
      * Address: 0x773616E4d11A78F511299002da57A0a94577F1f4
      */
     constructor() public {
-        priceFeed = AggregatorV3Interface(0x773616E4d11A78F511299002da57A0a94577F1f4);
+        priceFeed = AggregatorV3Interface(
+            0x773616E4d11A78F511299002da57A0a94577F1f4
+        );
     }
 
     /**
      * @return The price of DAI expressed in ETH.
      */
-    function consult() public view returns (int) {
+    function consult() public view returns (int256) {
         (
             uint80 roundID,
-            int price,
-            uint startedAt,
-            uint timeStamp,
+            int256 price,
+            uint256 startedAt,
+            uint256 timeStamp,
             uint80 answeredInRound
         ) = priceFeed.latestRoundData();
         return price;
